Finished

+++ src/controllers/to-do.controller.js
@@ -1,33 +1,86 @@
 const ToDoService = require('../services/to-do.service');
+const mongoose = require('mongoose');
 
-const findAllToDoController = (req, res) => {
-  const to_do = ToDoService.findAllToDoService();
-  res.send(to_do);
+
+const findAllToDoController = async (req, res) => {
+  const ToDo = await ToDoService.findAllToDoService();
+  res.send(ToDo);
 };
 
-const findByIdToDoController = (req, res) => {
-  const parametroId = Number(req.params.id);
-  const escolhaTo_Do = ToDoService.findByIdToDoService(parametroId);
-  res.send(escolhaTo_Do);
+const findByIdToDoController = async (req, res) => {
+  const idParam = req.params.id;
+
+  if (!mongoose.Types.ObjectId.isValid(idParam)) {
+    res
+      .status(400)
+      .send({ message: 'ID inválido!' });
+    return;
+  }
+
+  const escolhaToDo = await ToDoService.findByIdToDoService(idParam);
+
+  if (!escolhaToDo) {
+    return res.status(404).send({ message: "ToDo não encontrada!" })
+  }
+
+  res.send(escolhaToDo);
 };
 
-const createToDoController = (req, res) => {
-  const to_do = req.body;
-  const newToDo = ToDoService.createToDoService(to_do);
+const createToDoController = async (req, res) => {
+  const ToDo = req.body;
+  if (
+    !ToDo ||
+    !ToDo.time ||
+    !ToDo.description
+  ) {
+    res.status(400).send({ mensagem: "Você não preencheu todos os dados para adicionar uma nova ToDo a sua lista!" });
+
+  }
+  const newToDo = await ToDoService.createToDoService(ToDo);
   res.send(newToDo);
 };
 
-const updateToDoController = (req, res) => {
-  const idParam = Number(req.params.id);
-  const toDoEdit = req.body;
-  const updateToDo = ToDoService.updateToDoService(idParam, toDoEdit);
+const updateToDoController = async (req, res) => {
+  const idParam = req.params.id;
+  const ToDoEdit = req.body;
+
+  if (!mongoose.Types.ObjectId.isValid(idParam)) {
+    res.status(400).send({ message: 'ID inválido!' });
+    return;
+  }
+
+  const escolhaToDo = await ToDoService.findByIdToDoService(idParam);
+
+  if (!escolhaToDo) {
+    return res.status(404).send({ message: "ToDo não encontrada!" })
+  }
+
+
+  if (!ToDoEdit || !ToDoEdit.time || !ToDoEdit.description) {
+    return res.status(400).send({ message: "Você não preencheu todos os dados para editar a ToDo!" });
+  }
+
+  const updateToDo = await ToDoService.updateToDoService(idParam, ToDoEdit);
   res.send(updateToDo);
 };
 
-const deleteToDoController = (req, res) => {
-  const idParam = Number(req.params.id);
-  ToDoService.deleteToDoService(idParam);
-  res.send({message:'Item da lista deletado com sucesso!'})
+const deleteToDoController = async (req, res) => {
+  const idParam = req.params.id;
+
+  if (!mongoose.Types.ObjectId.isValid(idParam)) {
+    res.status(400).send({ message: 'ID inválido!' });
+    return;
+  }
+
+  const escolhaToDo = await ToDoService.findByIdToDoService(idParam);
+
+  if (!escolhaToDo) {
+    return res.status(404).send({ message: "ToDo não encontrada!" })
+  }
+
+
+  await ToDoService.deleteToDoService(idParam);
+  res.send({ message: 'Item da lista deletado com sucesso!' })
 };
 module.exports = {
   findAllToDoController,
